Revert "block: add proper state guards to __elv_next_request"
This reverts commit 07bdfb8d38a5c9aee4451c10c58302081d47733e.

It's a follow up fix to the commit "put stricter guards on queue dead
checks", which caused a regression reported in bug 793796, and thus was
reverted. But this follow up fix doesn't fix the crash reported at bug
793796 (confirmed by reporter on comment #37), so it is being reverted
as well.

The original regression report is against Natty, but Lucid inherited
same changes through stable updates, and has same issues, as also
reported and confirmed on bug 811745.

This revert is being tracked in bug 802986.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>