Revert "Fix oops caused by queue refcounting failure"
This reverts commit 93b83aacba2a69055ea0fd6984f73824b52f4147.

It's another follow up fix to the commit "put stricter guards on queue
dead checks", which caused a regression reported in bug 793796, and thus
was reverted. But this additional fix doesn't fix the crash reported at
bug 793796 (confirmed by reporter on comment #39), so it is being
reverted as well.

The original regression report is against Natty, but Lucid inherited
same changes through stable updates, and has same issues, as also
reported and confirmed on bug 811745.

This revert is being tracked in bug 802986.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>